Working on environment and also score system
Environment update as well as a first push of the scoring system and how the balls spawn. The OnTriggerEnter is still a little buggy and I will be working on it more tonight. Once that is fixed though everything should be good for the scoring system and with balls spawning too